+++ test/html/test_html.py
@@ -637,7 +637,16 @@ def test_html_blockquote_indent(tmp_path):  # issue-1074
     pdf = FPDF()
     pdf.add_page()
     html = "Text before<blockquote>foo</blockquote>Text afterwards"
-    pdf.write_html(html, tag_indents={"blockquote": 5})
+    pdf.write_html(html, tag_indents={"blockquote": 20})
+    html = (
+        "<blockquote>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod"
+        "tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,"
+        "quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."
+        "Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu"
+        "fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident,"
+        "sunt in culpa qui officia deserunt mollit anim id est laborum.</blockquote>"
+    )
+    pdf.write_html(html, tag_indents={"blockquote": 40})
     assert_pdf_equal(pdf, HERE / "html_blockquote_indent.pdf", tmp_path)
 
 
@@ -645,7 +654,9 @@ def test_html_li_tag_indent(tmp_path):
     pdf = FPDF()
     pdf.add_page()
     with pytest.warns(DeprecationWarning):
-        pdf.write_html("<ul><li>item</li></ul>", li_tag_indent=10)
+        pdf.write_html("<ul><li>item 1</li></ul>", li_tag_indent=40)
+        pdf.write_html("<ul><li>item 2</li></ul>", li_tag_indent=50)
+        pdf.write_html("<ul><li>item 3</li></ul>", li_tag_indent=60)
     assert_pdf_equal(pdf, HERE / "html_li_tag_indent.pdf", tmp_path)
 
 
@@ -686,3 +697,130 @@ def test_html_ol_ul_line_height(tmp_path):
     </ul>"""
     )
     assert_pdf_equal(pdf, HERE / "html_ol_ul_line_height.pdf", tmp_path)
+
+
+def test_html_long_list_entries(tmp_path):
+    pdf = FPDF()
+    pdf.add_page()
+    html = f"<ul><li>{'A ' * 200}</li></ul>"
+    pdf.write_html(html)
+    assert_pdf_equal(pdf, HERE / "html_long_list_entries.pdf", tmp_path)
+
+
+def test_html_long_ol_bullets(tmp_path):
+    pdf = FPDF()
+    pdf.add_page()
+    html_arabic_indian = f"""
+            <ol start="{10**100}">
+              <li>Item 1</li>
+              <li>Item 2</li>
+              <li>Item 3</li>
+            </ol>
+        """
+    html_roman = f"""
+            <ol start="{10**5}" type="i">
+              <li>Item 1</li>
+              <li>Item 2</li>
+              <li>Item 3</li>
+            </ol>
+        """
+    pdf.write_html(html_arabic_indian)
+    pdf.write_html(html_roman, type="i")
+    pdf.write_html(html_arabic_indian, tag_indents={"li": 50})
+    pdf.write_html(html_roman, tag_indents={"li": 100})
+    assert_pdf_equal(pdf, HERE / "html_long_ol_bullets.pdf", tmp_path)
+
+
+def test_html_measurement_units(tmp_path):
+    for unit in ["pt", "mm", "cm", "in"]:
+        pdf = FPDF(unit=unit)
+        pdf.add_page()
+        html = """
+                <ul>
+                    <li>Item 1</li>
+                    <li>Item 2</li>
+                    <li>Item 3</li>
+                </ul>
+                <ol>
+                    <li>Item 1</li>
+                    <li>Item 2</li>
+                    <li>Item 3</li>
+                </ol>
+                <blockquote>Blockquote text</blockquote>
+                <dt>Description title</dt>
+                <dd>Description details</dd>
+            """
+        pdf.write_html(html)
+        assert_pdf_equal(pdf, HERE / "html_measurement_units.pdf", tmp_path)
+
+
+def test_bulleted_paragraphs():
+    pdf = FPDF()
+    pdf.add_page()
+    text_columns = pdf.text_columns(skip_leading_spaces=True)
+    cases = [
+        {"indent": 1, "bullet_string": None},
+        {
+            "indent": 10,
+            "bullet_string": "",
+            "bullet_r_margin": 2,
+        },
+        {
+            "indent": 20,
+            "bullet_string": "a",
+            "bullet_r_margin": 0,
+        },
+        {
+            "indent": -20,
+            "bullet_string": "abcd",
+            "bullet_r_margin": 4,
+        },
+        {
+            "indent": 1000,
+            "bullet_string": "abcd\nfghi",
+            "bullet_r_margin": -3,
+        },
+    ]
+    pdf.set_font("helvetica", "B", 16)
+    for case in cases:
+        try:
+            text_columns.paragraph(
+                indent=case.get("indent"),
+                bullet_string=case.get("bullet_string"),
+                bullet_r_margin=case.get("bullet_r_margin"),
+            )
+            text_columns.end_paragraph()
+        except FPDFException as error:
+            pytest.fail(
+                f"case: (indent: {case['indent']}, bullet_string: {case['bullet']})\n"
+                + str(error)
+            )
+    bad_bullet_string = "我"
+    with pytest.raises(FPDFException) as error:
+        text_columns.paragraph(indent=1, bullet_string=bad_bullet_string)
+    expected_msg = (
+        f'Character "{bad_bullet_string}" at index {0} in text is outside the range of characters '
+        f'supported by the font used: "{pdf.font_family+pdf.font_style}". Please consider using a Unicode font.'
+    )
+    assert str(error.value) == expected_msg
+
+
+def test_html_list_vertical_margin(tmp_path):
+    pdf = FPDF()
+    for margin_value in (None, 4, 8, 16):
+        pdf.add_page()
+        html = f"""
+            This page uses `list_vertical_margin` value of {margin_value}
+             <ul>
+                <li>Item 1</li>
+                <li>Item 2</li>
+                <li>Item 3</li>
+            </ul>
+            <ol>
+                <li>Item 1</li>
+                <li>Item 2</li>
+                <li>Item 3</li>
+            </ol>
+        """
+        pdf.write_html(html, list_vertical_margin=margin_value)
+    assert_pdf_equal(pdf, HERE / "html_list_vertical_margin.pdf", tmp_path)